feat(web): add action button
