+++ apps/web/src/lib/icons/icon-arrow-right.svelte
@@ -1,0 +1,20 @@
+<script lang="ts">
+  import type { SVGAttributes } from 'svelte/elements'
+
+  type Props = SVGAttributes<SVGSVGElement>
+
+  let { ...rest }: Props = $props()
+</script>
+
+<svg
+  xmlns="http://www.w3.org/2000/svg"
+  width="1em"
+  height="1em"
+  viewBox="0 0 256 256"
+  fill="currentColor"
+  {...rest}
+>
+  <path
+    d="M221.66,133.66l-72,72a8,8,0,0,1-11.32-11.32L196.69,136H40a8,8,0,0,1,0-16H196.69L138.34,61.66a8,8,0,0,1,11.32-11.32l72,72A8,8,0,0,1,221.66,133.66Z"
+  />
+</svg>
